@@ -2,44 +2,44 @@ module github.com/isindir/git-get
 
 // UPDATE_HERE
 // https://go.dev/dl/
-go 1.24.3
+go 1.25.5
 
 require (
 	// https://github.com/fatih/color/releases
 	github.com/fatih/color v1.18.0
 	// https://github.com/google/go-github/releases
-	github.com/google/go-github/v72 v72.0.0
+	github.com/google/go-github/v81 v81.0.0
 	// https://github.com/ktrysmt/go-bitbucket/releases
-	github.com/ktrysmt/go-bitbucket v0.9.85
+	github.com/ktrysmt/go-bitbucket v0.9.88
 	// https://github.com/sirupsen/logrus/releases
 	github.com/sirupsen/logrus v1.9.3
 	// https://github.com/spf13/cobra/releases
-	github.com/spf13/cobra v1.9.1
+	github.com/spf13/cobra v1.10.2
 	// https://github.com/stretchr/testify/releases
-	github.com/stretchr/testify v1.10.0
-	// https://github.com/xanzy/go-gitlab/releases
-	github.com/xanzy/go-gitlab v0.115.0
+	github.com/stretchr/testify v1.11.1
+	// https://gitlab.com/gitlab-org/api/client-go/-/releases
+	gitlab.com/gitlab-org/api/client-go v1.11.0
 	// https://pkg.go.dev/golang.org/x/net
-	golang.org/x/net v0.40.0
+	golang.org/x/net v0.48.0 // indirect
 	// https://pkg.go.dev/golang.org/x/oauth2
-	golang.org/x/oauth2 v0.30.0
+	golang.org/x/oauth2 v0.34.0
 	// https://gopkg.in/yaml.v3
 	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/google/go-querystring v1.1.0 // indirect
+	github.com/google/go-querystring v1.2.0 // indirect
 	github.com/hashicorp/go-cleanhttp v0.5.2 // indirect
-	github.com/hashicorp/go-retryablehttp v0.7.7 // indirect
+	github.com/hashicorp/go-retryablehttp v0.7.8 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mitchellh/mapstructure v1.5.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/spf13/pflag v1.0.6 // indirect
+	github.com/spf13/pflag v1.0.9 // indirect
 	github.com/stretchr/objx v0.5.2 // indirect
-	golang.org/x/sys v0.33.0 // indirect
-	golang.org/x/time v0.3.0 // indirect
+	golang.org/x/sys v0.39.0 // indirect
+	golang.org/x/time v0.14.0 // indirect
 )
